pleioverse: fast self-contained pleiotropy simulation and testing

A high-performance C++/Rcpp toolkit for pleiotropic genetic data, unified around the pleio object. It provides run_pleiosim() to simulate pleiotropic multi-phenotype data, run_pleiotest() to run the full suite of multiple-association testing methods and pleiotropy tests, and one run_<test>() function per individual test.

See also

Author

Maintainer: Wanjun Gu wanjun.gu@ucsf.edu